shared/pyproject.toml
[project]
name = "shared"
version = "0.1.0"
description = "Shared utilities for multi-agent interoperability"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "a2a-sdk[all]",
  "openai-agents",
  "httpx",
  "opentelemetry-api",
  "opentelemetry-sdk",
  "opentelemetry-instrumentation-fastapi",
  "opentelemetry-instrumentation-httpx",
  "opentelemetry-instrumentation-logging",
  "opentelemetry-exporter-otlp-proto-grpc",
]

[tool.uv]
package = true

[dependency-groups]
dev = [
  "ruff",
]

[tool.ruff]
target-version = "py313"

[tool.ruff.lint]
select = ["ALL"]
ignore = [
  "CPY",
  "D417",
  "D418",
  "D419",
  "DOC",
]
